Fix some broken ZipFile unicode tests
These tests rely on the use of the FileData class to verify zips are created/extracted with the correct files. However, FileData and the accompanying helper methods do not take unicode normalization into account, so unicode zip tests are failing on OSX. I modified these tests to verify based on the actual files being zipped/unzipped instead of using the expectations set in the FileData class.