Avoid unnecessary clearing in CopyToAsync
With getting buffers from the ArrayPool, we want to clear the buffers before we put them back into the pool.  But we only need to clear the max amount we used, and we're very likely to be using much larger buffers than we actually need.  This change keeps track of how much space we used and only clears that much.