Use CommonCrypto CSPRNG for GetTempFileName on macOS
In the Unix build use the existing calculation of IsMac to determine if
Interop.Crypto (powered by OpenSSL) or CommonCrypto should be used
for the CSPRNG.

GetTempFileName is the most often reported hit of a failure in Interop+Crypto.cctor
, but that didn't benefit from when the macOS
implementation of System.Security.Cryptography.RandomNumberGenerator
changed to use CommonCrypto (via the new Apple-specific shim).

This rectifies that, and should take a significant burden off of the first run
experiences for macOS users who have not installed OpenSSL in a manner
where we can find it.