Build Linq.Expression tests against live Regex
The version for Regex was downgraded when I moved the API to netcoreapp
instead of netstandard.  This caused a failure in CI because the
Linq.Expression tests were building against packages (which had the
higher version) but running against live bits; due to
"TestWithLocalLibraries" used in the CI.

Temporarily build this test against the live bits.  This can be removed
with the next package update.