Add WebClient to System.Net.Requests
This PR ports System.Net.WebClient to corefx.  The code is mostly taken from desktop and then cleaned up a bit stylistically.  The only major code rewrite was removing hundreds of lines of complicated APM callback-based code and replacing it with a few core async methods.  There's still plenty of more cleanup that can be done, but functionally this is sufficient. I also wrote a bunch of tests for it (~85% coverage).