Fix WinHttpHandler when connecting thru authenticating proxy
WinHttpRequestCallback.OnRequestSendingRequest was throwing ERROR_WINHTTP_INCORRECT_HANDLE_STATE when trying
to retrieve the WINHTTP_OPTION_SERVER_CERT_CONTEXT during custom server certificate validation. This problem
would repro when POST'ing a request body to a secure (TLS/SSL) server thru an authenticating proxy.

The error was being returned because the request handle was not yet bound to a secure connection. This occurs
while connecting thru a proxy where the CONNECT verb hasn't yet been processed due to the proxy requiring
authentication. Discussions with the WinHTTP team confirmed this analysis and fix.

The fix is to ignore this notification when getting this error. Another notification will be sent once the final
secure connection thru the proxy is completed.

Fixes #11973