Fixed performance of Min/Max in SortedSet.
The existing implementation of Min/Max calls InOrderWalkTree,
which perform an allocation of a stack for traversing the tree.
But when you are finding the smallest or largest element in a
tree there is no need to perform this allocation.  This is an
optimised version that directly finds the smallest or largest
element.