[Fixed test] Wait for process to exit before asserting HasExited
Kill() is async, so we need to wait for the process to actually exit
before asserting HasExited. Might as well merge the new test with the
existing one which already does WaitForExit. Kept separate asserts to
make it easier to investigate test failures. I couldn't repro locally,
but this should work (thank you CI). If it doesn't, I'll work harder on
a repro.

Validated that the test fails without the Process.HasExited change, and
succeeds with the change (ran -x500).

[Please squash with initial commit]