Always define GeneratedProjectJsonDir property
Fixes https://github.com/dotnet/corefx/issues/12001

Apparently msbuild will try to expand the wildcard item even
if the condition on it is false and in some cases that causes
GeneratedProjectJsonDir to be set to empty and causes msbuild
to try to find all project.json files on the entire drive. To ensure
we don't try indexing the entire drive we will always set
the property.

https://github.com/dotnet/corefx/commit/760104f2957978c475648923f744f0b709507559#diff-0b192804a6349e8c26d2b027afbd89a2R131