Fix NoSocketsReuseUnicastPortSupport tests
The NoSocketsReuseUnicastPortSupport tests were failing on Windows Server 2012 R2 in the CI runs because we thought the feature was not present. It was only introduced in Windows 10. However, it was later backported as a serving fix to both Windows 8.1 (Windows Server 2012 R2) and Windows 8 (Windows Server 2012). It is not possible to detect in the CI tests if a servicing patch is installed. So, this PR adjusts the capability detection to return a null value (instead of true or false) to indicate this uncertainty and thus the tests are skipped in that case.

Fixes #11088