Expose path constructors for StreamReader/StreamWriter
Exposes the members and adds new tests for them. System.IO is already updated to build for netstandard1.7 so no changes to be made there. These new constructors are not added for uap10.1 as that would require we merge the IO and IO.FileSystem assemblies to gain access to FileStream from the IO assembly.