Always initialize Targets* properties to false
We were hitting some cases where TargetsWindows was set as and
environment variable and caused some of our builds to start failing.

These should never be set outside of the Choose statement based on
OSGroup so we are always defaulting them to false to ensure the enviroment
doesn't end up overriding them.