Fix Uri handling of unknown schemes
Somehow this expression got transposed in corefx; it should be ```end - startOtherHost``` rather than ```startOtherHost - end``` (as can be seen in the previous section of code, end will always be >= startOtherHost).  As it was, it was resulting in an ArgumentOutOfRangeException, which a test was then codifying, and that test was then failing on desktop which has the correct ordering.