Fix XmlDocument target framework and XPath assembly version
XmlDocument should have been moved to NS17. By not doing so and
removing forwarding all the types to S.XML.ReaderWriter in NS13
is a breaking change.

XPath has added new APIs but the Assembly Version wasn't bumped
so moved it to 4.1.0

S.Xml.Xsl.Primitives is a dead contract that has never shipped and
is now folded int S.Xml.ReaderWriter so I deleted it.

Cleaned up a bunch of project.json files to elimate build warnings
and make them consistent, focused on Xml libraries but also fixed
a few others to eliminate the build warnings.