Fix XmlDocument target framework and XPath assembly version (#12185)
XmlDocument should have been moved to NS17. It is a breaking
change to type-forward all the types to S.Xml.RewaderWriter
in NS13 because those types don't exist in the NS13 version
of that library.

XPath has added new APIs but the Assembly Version wasn't bumped
so moved it to 4.1.0

S.Xml.Xsl.Primitives is a dead contract that has never shipped and
is now folded into S.Xml.ReaderWriter so I deleted it.

Cleaned up a bunch of project.json files to eliminate build warnings
and make them consistent. Focused on Xml libraries but also fixed
a few others to eliminate the build warnings.