Make PKGProj's restore the baseline project.json
PKGProj build depends on harvesting dlls from previously shipped
packages.  Rather than force every package in the repo to declare the
version of the last-shipped package in project.json we have a central
copy listing everything.  The problem here is that if this project isn't
restored we'll just silently drop assets from the package, which can
cause tests to fail or package build to fail validation.  This also
causes a problem for folks that want to just sync the repo and build
a single project without doing a full restore.

To ensure that we'll always have the baseline restored I set this as the
ProjectJson for the pkgproj and ensure that we run restore during
package build.