Add nestandard1.7 build for Diagnostics.Debug
This retargets the ref and src to netstandard1.7.

The test now only targets nestandard1.7.  This is due to the test
actually *not* being a portable test.  It compiles against
implementation and won't really run on anything but our impl.

This test should probably be split between unit and conformance
where the conformance cases target netstandard and run anywhere
and unit cases can compile against implementation.