Workaround TFS sni.dll package mismatch
The sni.dll package dependencies were updated to 4.4 too soon, we
haven't yet gotten an updated build from TFS.

Until we do, make sure to use the version that matches what we're
getting from TFS: 4.3.0.  Once we get an update this can be reverted.

Workaround the mismatch in the current builds by downgrading to
the package versions that exist:
    "runtime.win10-arm64.runtime.native.System.Data.SqlClient.sni":
"4.3.0-beta-24523-00",
    "runtime.win7-x64.runtime.native.System.Data.SqlClient.sni":
"4.3.0-beta-24523-00",
    "runtime.win7-x86.runtime.native.System.Data.SqlClient.sni":
"4.3.0-beta-24523-00",