Workaround lack of Uri support for Unix paths in XML
```new Uri(filePath)``` currently works on Windows and fails on Unix; there are special code paths that guess whether the specified url is a file path and automatically prepend a "file://" scheme if it is, but the logic is based on Windows paths.  On Unix, it ends up throwing a UriFormatException.

A bunch of the Xml types end up hitting this code path when provided with a string uri, e.g. XmlReader.Create(string), which causes failures when file paths are provided on Unix, and creating readers and such for file paths is very common.  This commit works around the issue by adding an explicit file scheme rather than leaving it up to the Uri implementation to determine that it's a file path.