Fix System.IO build with ArrayPool dependency
The previous commit added usage of StreamHelpers.cs.  The dependency is only on some validation routines, but the same file also includes usage of ArrayPool, and thus the build breaks without a dependency on System.Buffers.  Since we don't currently have another need for System.Buffers in System.IO (this could change in the future), for now I'm splitting the StreamHelpers.cs file into two files, one that needs ArrayPool, and one that doesn't, so that System.IO can include only the latter.