Add some "manual" System.Console tests
Today our test suite for System.Console is significantly lacking, due to our current inability to handle reading, verification of output to the terminal, etc.  Until we are better at such things, this PR adds a manual test suite, that instructs the user to do something and to then verify specific visual results and behaviors.