Fix Console.CursorLeft/Top handling in WSL
Console.get_CursorLeft/Top work by issuing a request to the terminal by writing out a known sequence to stdout.  The terminal then provides a response on stdin, again in a known format: \ESC[row;colR.  Our current logic reads characters until it finds the escape, then reads until it finds the brace, and then reads until it gets a digit that starts the row.  At that point, it parses whatever value it has in its input buffer and treats that as the row.  Then it does the same for the column.  This has worked fine on all of the systems we tested on, but it breaks in the Windows Subsystem for Linux.  The reason is that the other systems all apparently write the response atomically, such that when we go to read the data, all of the digits for the row/column are available.  But WSL writes each digit out individually, which means when we read the first digit, there's a good chance the remaining digits did not get read with it into our buffer, and we end up misinterpreting just that first digit as being the entire value.  So for example, if WSL responds to a request with \ESC[12;34R, we're likely to interpret the current position as being 1,3 rather than 12,34.  This commit fixes this by reading each character individually and not declaring that we're done parsing the row or column until we hit the known marker that comes after it, namely the ';' in the case of the row and the 'R' in the case of the col.

(Note that I got rid of the extra buffering we were doing here, as it was flawed and would have added additional code to do it correctly with this new approach.  If we determine that there's a noticeable performance impact to not having it, then we can add it in.)