Update DbCommand.cs
Since base implementation of the asynchronous methods is provided, i think that they should at least run in a worker thread so that they won't block. Drivers that extend this class could then override the asynchronous methods and implement them in their own way (probably with Asynchronous I/O).

The MySQL Driver for instance doesn't override any of the asynchronous methods in this class so every attempt to run a query asynchronous ends up with a blocking code.