Avoid using DefaultThreadCurrentCulture in tests
We still have some tests that use DefaultThreadCurrent{UI}Culture in the current process.  All usage of DefaultThreadCurrent{UI}Culture needs to either be done in a separate process or changed to Current{UI}Culture, as the former affects not only the current thread, but also the default used for any new threads created in the process, which leads to non-deterministic failures in our test suites.

This PR fixes all remaining usage in our existing tests.