Fix DateTimeFormatInfoGetInstance.GetInstance test
The test was using Assert.Equal to compare the DateTimeFormatInfo returned from GetInstance with the one provided by the theory as the expected result.  Two problems with this: 1) DateTimeFormatInfo doesn't override equality testing, so Equal will end up doing an object reference comparison, and 2) xunit doesn't guarantee that the thread which enumerates a member data is the same that invokes all instances of the associated test.  As a result, we could end up in cases where the member data happened to be enumerated on one thread, the test was invoked on another, and the identity of DateTimeFormatInfo.CurrentInfo would then not match, resulting in equality failures.