Remove defunct AssertWithCallerAttributes from tests
When we didn't have file name / line number information in exception stack traces, we added this AssertWithCallerAttributes.cs file, which could be included in a test project to force asserts to bind to it instead of to those from xunit, and it uses caller attributes to include such information in the exceptions.  Now that we have appropriate information provided by the runtime, we no longer need this workaround.