Enable support of ephemeral certificates with EnvelopedCms
EnvelopedCms resolves the key objects from a certificate via P/Invokes instead
of calling public API (like GetRSAPrivateKey()).  The code that it has is not
capable utilizing ephemeral keys (imported with PKCS12_NO_PERSIST_KEY).

This change makes it so that ephemeral keys imported into CNG will work,
meaning that certificates opened from a PFX with
`X509KeyStorageFlags.EphemeralKeySet` will function. (Ephemeral keys imported
into CAPI are not supported with this change, as they're stored differently in the
PCERT_CONTEXT)

The tests will be as ephemeral as possible (perphemeral when running as
netstandard1.3, ephemeral when running as netcoreapp1.1). Since perphemal
is still needed for the default decryption tests they're staying marked as
OuterLoop.