Another fix for the TickCount test.
From test failure logs, it appears that this test tends to fail when the
"Spin block" test is only performed once. This seems to indicate that the
TickCount is only read once upon entering the spin. The SpinUntil call
terminates (after one second) before the TickCount is ever read again. If
we encounter a scenario like that, we now attempt to read the TickCount once
more after the SpinUntil call, to ensure we have read the TickCount after
the second delay.