Fix parsing of procfs stat files when comm name contains spaces
If a process or thread name contains a space, our procfs stat file parsing logic breaks as it sees the space in the name as a separator between elements.  This in turn causes Process.GetProcesses() and other such functions that rely on these stat files to fail with InvalidDataExceptions.  This commit fixes it by adding knowledge of '(' and ')' groupings to the parser.