Make Microsoft.CSharp target netstandard1.7
Makes the ref, src, & tests target netstandard1.7.
The tests cross-compile to preserve the test build for older fxs.

I intentionally *did not* bump the version of this assembly.  The reason
being that a bump would require an update to desktop to ship this
assembly out of band, however that isn't needed.  Since all the new API
is already part of desktop we can make this work while still treating it
as inbox.