Stop crashing in NegotiateStream tests on Unix when KDCSetup fails
The NegotiateStream tests on Unix try to set up kerberos infrastructure.  The tests then check whether infrastructure is available, and gracefully don't run if it's not.  But the KDCSetup fixture is throwing an exception if the infrastructure can't be configured, which is causing a fair number of our outerloop CI runs to fail on various OSes.  Until we have a better solution in place, I'm making the failure graceful rather than crashing.