Change the cryptography shims to use direct package dependencies
The use of runtime dependencies for the platform-specific native shims causes
some upgrade complications.  The design of the shared framework results in
the dependency being considered to be satisfied even when a higher version
was needed.  So if someone has .NET Core 1.0 installed and pulls in a daily build
of a cryptography library they might be missing native code changes that were
assumed to be present as dependencies of the managed code.

By changing to direct dependencies (changing the platform-specific package
files to use IdPrefix instead of PackageTargetRuntime) the new .so files will be
downloaded and deployed with the app (if they're a higher version that what
is present in the shared framework).