Fix binclash in NetworkInfo tests
I had changed the builds file to build a configuration of the tests that
didn't use an OSGroup (since the tests don't need to compile OSGroup-
specific).  The problem was the tests were still defaulting OSGroup via
a configuration string, so the build with and without the OSGroup had
the same TargetPath and would clash/race in a multi-core build.