Update all projects to last stable packages
And remove imports.

Now that we have netstandard packages that support all targets we should
use them.  This means we don't need imports.

I opted to use the stable packages rather than the LKG ones since this
will reduce churn in the product assemblies.

This has the added benefit of reducing download time for the repo.