Add System.Security.Claims serialization
The serialization implementation is a limited port of what's in desktop, as for now at least we want to avoid a dependency on BinaryFormatter, and without significant refactoring we can't take a dependency on WindowsIdentity, which sits in a higher-level contract.  For example, attempting to serialize a ClaimsPrincipal containing any identities will throw a PlatformNotSupportedException.

Note that I removed the test configuration to support down-level, as there are essentially 0 existing tests... the only one here is one I previously added so that there would be > 0 tests so that it would show up in code coverage, highlighting that there were zero tests :)