Fix ValueTask behavior as async return type
There's a bug in ValueTask<T>'s builder based on a misunderstanding of what the codegen would be from the compiler.  I'd thought the compiler would codegen access to the builder's Task property, but that was based on a very old model; the model actualy in place relies on the Await*OnCompleted methods doing any initialization that's necessary prior to the await.  That's an important distinction, because the builder for ValueTask needs defer to the wrapped task builder's Task property once any await has been issued.  Which means that rather than just delegating to the wrapped builder's Await*OnCompleted methods, before doing so the ValueTask's builder must remember that these methods have been called, such that a subsequent access will defer to the wrapped Task, even if a result is set in the interim.