Fix null reference exception in NetworkAvailabilityChanged on Linux
If NetworkAddressChanged has a subscriber and NetworkAvailabilityChanged doesn't, attempting to remove a delegate from NetworkAvailabilityChanged results in a NullReferenceException as a null timer static field is dereferenced.

This was failing intermittently in tests because the NetworkAddressChanged and NetworkAvailabilityChanged tests were in different classes and thus could potentially run in parallel.  If the NetworkAvailabilityChanged_JustRemove_Success test ran while one of the NetworkAddressChanged tests was in flight, it would fail with a null ref.

I've fixed the product bug by adding a null check.  I've also updated the tests so that all of these event handler tests are in the same class, and added explicit tests for this case that will deterministically fail if the product regresses.