Reduce Process.Modules allocations
The implementation of `Process.Modules` builds up an array of
intermediate `ModuleInfo` objects, which are then wrapped as
`ProcessModule` objects and copied to a new `ProcessModuleCollection`.
These intermediate object/array allocations can be avoided by deleting
the `ModuleInfo` concept and just using `ProcessModuleCollection` to
build up the collection of `ProcessModule` objects.

The implementation on Windows can be further improved by avoiding some
more unnecessary allocations, such as using a single `char[]` instead
of duplicate `StringBuilder` instances to retrieve each module's
`ModuleName` and `FileName`, and changing `NtModuleInfo` to be a struct
instead of a class.