Enable build tests against packages as the default testing scenario
Currently we build and test in CI differently from the official pipeline. This change will help reduce inconsistencies in the way in which we build to expose issues sooner.
The build will now build the product as packages, install them locally, restore the test-runtime deps for the locally built versions and then test them. The locally built
packages are installed to the localpackages folder and the LKG set restored goes into the packages folder. Tests will only restore against locally built packages as a
source and then copy the required LKG dependencies like specific runtime dependencies that aren't built locally. The localpackages are overwritten each time a build occurs so that only one version remains in the folder.

Other changes required to make this work:
 - Add missing runtimes in the supports clause
 - Add LKG runtime packages not produced by the build