Fix TcpClient.BeginConnect(IPAddress, ...) on Unix
Fixing the issue that caused the sockets tests not run on Unix then revealed a test failure on Unix with TcpClient.BeginConnect.  It needs to use TaskToApm.Begin, as the EndConnect method expects to be able to use TaskToApm.End.  This not being done correctly was resulting in an ArgumentNullException prior to the fix.