Fix the P/Invoke target for System.Security.Cryptography.Algorithms.Tests
The crypto shim got renamed, but a capabilities probe in the S.S.C.Algorithms
test library didn't get the memo.  This didn't matter much because the old shim
name is still produced for compatibility, but it got highlighted in a test with
packages change.

This change removes the duplicate definition of the library name, and includes
the Unix library names list, but leaves the individual targeted methods in place.