Add deserialization ctors where needed in permissions
Also added a test that serializes/deserializes an instance of every type in the assembly that's marked as serializable, just to make sure no exceptions are thrown.  In the process of doing so, I ran into an issue where xunit was trying to enumerate some of the permissions collections objects, which were returning null from GetEnumerator, which LINQ (used by xunit) does not expect, so I made those return empty enumerators rather than returning null.