Make TargetGroup-empty be for netcoreapp
This refactors all builds to make the netcoreapp-latest to be the
default build.

To do this I needed to be specific about the contract project in a few
cases so I made a fix to build tools to permit that and brought over
that fix as part of this commit.