Cleanup include: compile usage in test-runtime
This was being used for packages that were platform-specific.

Instead of using this hack don't include those in the netstandard
dependency sections.  Make sure we do include them (without using
the `"include": "compile"` constraint) in the frameworks that do support
them.

Make a netcoreapp1.0 specific section that includes Loader, otherwise
this is just a copy of netstandard1.6 section which was previously being
used.