Fix some race conditions in TypeConverter lib
There are a whole bunch of potential race conditions in the TypeConverter library.  This does not attempt to fix all of them, just address the symptoms of the ones that have been causing some failed tests.

In particular, the static TypeDescriptor.GetProperties will return a cached collection of instances, which means that instances end up being shared between threads.  Each ReflectPropertyDescriptor instance has properties that check whether a bit flag contains a bit indicating that the value has been lazily created, and if it doesn't, sets the bit, then creates and stores the value.  This leads to a variety of race conditions, including two threads racing to access the same value and one of them getting back the uninitialized value (this can also happen if an exception occurs in one thread before updating the value, leaving the flag set but the value uninitialized), and including two threads racing to access different fields but cloberring each other's updates to the state flags.

This PR addresses those two issues specifically (there are very likely others), by a) setting the bit only after the fields have been set, and b) using compare/exchange operations to update the set bits. The desktop code has the same issues, and I'm just making enough changes to avoid the test failures we've been seeing that arise due to this.