Eliminate SecBufferDesc allocations in System.Net.Security
Each SSPI operation in System.Net.Security allocates a SecBufferDesc to pass down to native code.  This change changes it from a class to a struct to avoid the heap allocation on each operation.  As the data is passed around, it's now passed by ref, except for the few code paths where we need to be able to pass null to the native code, in which case a ```SecBufferDesc*``` is used instead of a ```ref SecBufferDesc```.