Correctly encode the 0-subidentifier in DerEncoder
During the ECDSA-on-AppleCrypto work it was observed that two of the three
supported curves didn't import successfully due to the curve OIDs being
incorrect. These curves both had a sub-identifier with value 0, which was not
correctly handled by the DerEncoder class.

By changing while (!=0) to do..while(!=0) the 0-subidentifier gets a zero byte
pushed onto the data stack, making it encode as { 0x00 } instead of { }.

Added tests for both reading and writing.