Change NetworkStream to use base.ReadAsync/WriteAsync
The base Stream has an optimized implementation for delegating to Begin/EndRead.  Now that NetworkStream overrides those, we should use the base stream's implementation rather than using FromAsync (we can put back a custom implementation if/when we can do better than the base).  As the overrides exist in the contract, I've left the overrides in the implementation and just delegate directly to the base.