Adjusts ticks to support positive UTC offsets
HttpRequestCachePolicy internally stores cacheSyncTime as UTC and convert it
as local time before returning (which may leads to lossy transformation)

Its testcase currently uses 42 tick as an input, which is valid only when
UTC offset is negative (42 tick is usually lost during conversion if UTC
offset is positive), which leads to #13462.

This commit uses 504000000000 (14 * 60 * 60 * 10^7) as an input to fix #13462.