Add more static DSA tests, including NIST CAVP test vectors.
The portion of the tests in the existing file were signatures generated by .NET
Framework's DSACng, which were added first during this exercise.

The NIST cases came as a followup.  The general flow for NIST cases was:
* Take one representative sample from legacy DSA (FIPS 186-2).
* For each (L,N,H) tuple for FIPS 186-4 take the first test case.
* If any of the further test cases looked interesting (such as having a leading 0x00),
include such test case.

N=224 is not supported by Windows, and Windows CNG is the only current
DSA implementation that supports FIPS 186-3, so they were omitted as not
worth the hassle yet of making a new test capability that was never met.

H=SHA224 is not supported by .NET, so none of those cases were included.