Enable compat shims/facades building
This adds the building of netstandard.dll facade
as well as the net461 set of shim facades along
with some APICompat checks.

To accomplish this we are dependent on a flat layout
that we can use for seeds so we are going to use the
layout that we currently produce for indexing.

The APICompat check is split between net461 and netstandard20 which
will allow us to more easily track the remaining work for
netstandard20.