Add safemath.h and debugmacrosext.h from coreclr repo
We'd like to be able to use builtins for overflow checking, but the relevant builtins (e.g. __builtin_mul_overflow) aren't available with the compiler versions we currently use on all platforms.  In the meantime, we'll use the ClrSafeInt implementation from coreclr.

This commit brings over the files, but they don't yet compile.