Add a reusable awaitable data structure
Like TaskCompletionSource, this RendezvousAwaitable allows for a result or error to be passed from a producer to a consumer, with a consumer awaiting the object and a producer setting the result/error into it.  It may be reused any number of times (serially), avoiding having to allocate a new TaskCompletionSource/Task for each operation.