Avoid allocating the CancellationTokenSource when not needed
A new CancellationTokenSource is currently always allocated, which then also means that the data structures inside the CTS that are lazily-initialized on first use will end up getting allocated as well for each operation.  But if the supplied user's token isn't cancelable and there's no timeout, we can avoid these allocations.