Significantly reduce Task-based allocations in request lifecycle
New TaskCompletionSource/Task objects are being used for every stage of the lifecycle.  We can instead reuse a single awaitable object to avoid most of those costs.  We can then also add a CopyToAsync implementation to the response stream, which further reduces those costs (CopyToAsync is used by HttpContent, so benefits here accrue not only to custom user code but also the rest of the implementation).