Avoid allocation(s) in ReadAsStreamAsync
We were using an async method to store and return the results from CreateContentReadStreamAsync, putting the resulting Stream into a field typed as Stream.  This required having an async method to unwrap the result of of CreateContentReadStreamAsync, which in the best case (when CreateContentReadStreamAsync completes synchronously) would result in just a Task<Stream> allocation, and in the worst case (when CreatecontentReadStreamAsync completes asynchronously) would result in allocations for that async method's machinery.  We can instead just store the Task<Stream> in the field instead of the Stream, and avoid those allocations entirely.  (This also avoids an extra Task allocation each time the same stream is retrieved, though that's not particularly common.)